fix hilt error
