+++ build.gradle.kts
@@ -2,10 +2,6 @@
 plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.kotlin.android) apply false
-}
-
-buildscript {
-    dependencies {
-        classpath(libs.hilt.android.gradle.plugin)
-    }
+    alias(libs.plugins.hilt.android.gradle.plugin) apply false
+    alias(libs.plugins.ksp) apply false
 }
